Adding abstract class
